requirements.txt
flask==2.0.2
mistune==0.8.4
rich==10.12.0
python-dotenv==0.19.1
argparse==1.4.0
